Fingerprint Problem solved
Before processing begins, each CLS Stored Procedure and View is evaluated to make sure that the latest and greatest code is being run.  A "fingerprint" value is calculated and compared to the value stored in the database. If they match, processing proceeds.  If not, processing is halted and the job fails.

When promoting code, the code must be saved, reloaded from the file and compiled, then the fingerprint needs to be calculated.   After calculation, the UPDATE statement can be modified to that the proper value is stored.  If the calculation is not the penultimate operation in the sequence, we risk the danger of the wrong value being stored, because the fingerprint value can change upon recompilation.

I was doing it in the wrong order.